requirements.txt
# Core
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# Transformers (BioBERT, CLIP, pipelines)
transformers>=4.42.0
sentencepiece>=0.1.99

# LangChain & LangGraph
langchain-core>=0.2.0
langchain-community>=0.2.0
langgraph>=0.1.0

# FAISS (for vector store)
faiss-cpu>=1.7.4

# Documents, PDFs, OCR
PyPDF2>=3.0.0
pdf2image>=1.17.0
pytesseract>=0.3.10

# Imaging
Pillow>=10.0.0

# Streamlit app
streamlit>=1.37.0

# Environment
python-dotenv>=1.0.0
